通知アニメーション

+++ DxlibObjectiveGameLauncher/NoticeCanvas.cpp
@@ -18,7 +18,7 @@ NoticeCanvas::NoticeCanvas()
 void NoticeCanvas::UpdateAction()
 {
 	for (auto it = notices.begin(); it != notices.end(); ) {
-		if ((*it)->GetDeletedFrame()->GetMouseSelected() || (*it)->GetIsNoLife()) {
+		if (((*it)->GetDeletedFrame()->GetMouseSelected() || ((*it)->GetIsNoLife() && (int)(*it)->GetNoticeFrame()->GetPos().y == (int)(*it)->GetNoticeInitializePos().y)) && !(*it)->GetNoticeFrame()->GetMouseSelected()) {
 			delete (*it)->GetDeletedFrame();
 			delete (*it)->GetNoticeFrame();
 			layer.DeleteObject((*it)->GetDeletedFrame());
@@ -31,9 +31,14 @@ void NoticeCanvas::UpdateAction()
 	}
 
 	int i = 0;
+	//clsDx();
 	for (auto& item : notices) {
-		item->GetNoticeFrame()->ChangeValueWithAnimation(&item->GetNoticeFrame()->GetVectorPointer(VectorType::POS)->y, notices[0]->GetNoticeFrame()->GetPos().y - item->GetNoticeFrame()->GetSize().y * (float)i, 1.f);
+		item->GetNoticeFrame()->ChangeValueWithAnimation(&item->GetNoticeFrame()->GetVectorPointer(VectorType::POS)->y, item->GetNoticeInitializePos().y - item->GetNoticeFrame()->GetSize().y * (float)i, 1.f);
+		item->GetDeletedFrame()->ChangeValueWithAnimation(&item->GetDeletedFrame()->GetVectorPointer(VectorType::POS)->y, item->GetNoticeInitializePos().y - item->GetNoticeFrame()->GetSize().y * (float)i, 1.f);
+		item->GetTextObject()->ChangeValueWithAnimation(&item->GetTextObject()->GetVectorPointer(VectorType::POS)->y, item->GetNoticeInitializePos().y + item->GetTextOffset().y - item->GetNoticeFrame()->GetSize().y * (float)i, 1.f);
+		item->GetTextObject()->CalcPos();
 		item->TimerUpdate();
+		//printfDx("%d, %.2f, %.2f\n", i, item->GetNoticeFrame()->GetPos().y, notices[0]->GetNoticeFrame()->GetPos().y - item->GetNoticeFrame()->GetSize().y * (float)i);
 		i++;
 	}
 }
